Integrate pyroute2-based interface code.
For some reason we don't receive RTM_NEWADDR for IPv6.  Not sure what
that's about, we receive them for IPv4 and we receive RTM_DELADDR for
both.  If absolutely necessary we can hack something where we listen
for route change messages and use them as a trigger to re-scan the
interfaces, but that's pretty sick.  Press on IPv4-only for now.